+++ examples/zip_array_write_read.rs
@@ -8,13 +8,13 @@ use std::{
 use zarrs::{
     array::{Array, ZARR_NAN_F32},
     array_subset::ArraySubset,
-    storage::{ReadableStorageTraits, ReadableWritableStorageTraits},
+    storage::{ReadableStorageTraits, WritableStorageTraits},
 };
 
 // const ARRAY_PATH: &'static str = "/array";
 const ARRAY_PATH: &str = "/";
 
-fn write_array_to_storage<TStorage: ReadableWritableStorageTraits>(
+fn write_array_to_storage<TStorage: WritableStorageTraits>(
     storage: Arc<TStorage>,
 ) -> Result<Array<TStorage>, Box<dyn std::error::Error>> {
     use zarrs::array::{chunk_grid::ChunkGridTraits, codec, DataType, FillValue};