+++ src/storage/storage_transformer/performance_metrics.rs
@@ -5,9 +5,8 @@ use crate::{
     metadata::Metadata,
     storage::{
         ListableStorage, ListableStorageTraits, ReadableListableStorage, ReadableStorage,
-        ReadableStorageTraits, ReadableWritableStorage, StorageError, StoreKey, StoreKeyRange,
-        StoreKeyStartValue, StoreKeys, StoreKeysPrefixes, StorePrefix, WritableStorage,
-        WritableStorageTraits,
+        ReadableStorageTraits, StorageError, StoreKey, StoreKeyRange, StoreKeyStartValue,
+        StoreKeys, StoreKeysPrefixes, StorePrefix, WritableStorage, WritableStorageTraits,
     },
 };
 use std::sync::{
@@ -87,13 +86,6 @@ impl StorageTransformerExtension for PerformanceMetricsStorageTransformer {
         &'a self,
         storage: ListableStorage<'a>,
     ) -> ListableStorage<'a> {
-        self.create_transformer(storage)
-    }
-
-    fn create_readable_writable_transformer<'a>(
-        &'a self,
-        storage: ReadableWritableStorage<'a>,
-    ) -> ReadableWritableStorage<'a> {
         self.create_transformer(storage)
     }